.migration-backup/requirements-crypto.txt
# Crypto/Web3 Trading AI - Additional Requirements

# ===== Web3 & Blockchain =====
web3>=6.0.0
eth-account>=0.10.0
eth-abi>=4.0.0
eth-keys>=0.4.0
eth-utils>=3.0.0
py-solc-x>=1.1.0

# Solana
solana>=0.30.0
solders>=0.18.0
anchorpy>=0.18.0

# ===== Exchange APIs =====
python-binance>=1.0.17
ccxt>=4.0.0

# ===== Async & Networking =====
aiohttp>=3.8.0
websockets>=11.0
asyncio>=3.4.3
uvloop>=0.17.0

# ===== Data & Analytics =====
requests>=2.31.0
pandas>=2.0.0
numpy>=1.24.0

# ===== Databases =====
redis>=4.5.0
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0

# ===== Machine Learning =====
scikit-learn>=1.3.0
tensorflow>=2.16.0
torch>=2.0.0
transformers>=4.28.0

# ===== Monitoring & Alerts =====
prometheus-client>=0.16.0
sentry-sdk>=1.19.0

# ===== Utilities =====
python-dotenv>=1.0.0
pyyaml>=6.0
loguru>=0.7.0
schedule>=1.2.0
python-dateutil>=2.8.2

# ===== Testing =====
pytest>=7.4.0
pytest-asyncio>=0.21.0

# ===== Existing requirements (from requirements.txt) =====
# These are already in requirements.txt but listed for completeness
# numpy>=1.24.0
# pandas>=2.0.0
# scikit-learn>=1.3.0
# matplotlib>=3.7.0
# yfinance>=0.2.0
# joblib>=1.3.0
# schedule>=1.2.0
# TA-Lib>=0.4.28
# python-dotenv>=1.0.0
# loguru>=0.7.0
# pytest>=7.4.0
# requests>=2.31.0
# tensorflow>=2.16.0
# seaborn>=0.12.0
# plotly>=5.15.0
# alpaca-trade-api>=3.0.0
# beautifulsoup4>=4.12.0
# praw>=7.7.0
# vaderSentiment>=3.3.2
# newsapi-python>=0.2.7
# web3>=6.15.0
# eth-account>=0.10.0
# eth-keys>=0.4.0
# eth-utils>=3.0.0
# streamlit>=1.28.0
# plotly>=5.15.0
# newsapi-python>=0.2.7
# praw>=7.7.1
# fredapi>=0.5.1
# python-dateutil>=2.8.2
